examples/qwen_ner/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
datasets==3.0.2
dill==0.3.8
docstring_parser==0.16
einop==0.0.1
einops==0.8.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.3
peft==0.13.2
pillow==10.2.0
propcache==0.2.0
psutil==6.1.0
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
safetensors==0.4.5
shtab==1.7.1
six==1.16.0
sympy==1.13.1
tokenizers==0.20.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.5
transformers==4.45.2
triton==2.3.0
trl==0.11.4
typing_extensions==4.9.0
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.16.0
